Cloud Provider Shell/src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-cp-core>=1.0.0,<1.1.0
cloudshell-shell-connectivity-flow>=1.0,<2.0
cloudshell-networking-juniper>=5.0,<6
cloudshell-shell-standards>=1.0,<2.0
pyyaml